# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Casa Natura API",
    "version": "1.0.0",
    "description": "Full API documentation for Casa Natura (public + admin endpoints).",
    "contact": {
      "name": "Casa Natura",
      "url": "https://www.deltagreen.it"
    }
  },
  "servers": [
    { "url": "https://www.deltagreen.it", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "auth_token" }
    },
...
  },
  "paths": {
...
    "/api/cart/add": {
      "post": {
        "summary": "Add item to cart",
        "tags": ["Cart"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["productId", "variantId"], "properties": { "productId": {"type": "string"}, "variantId": {"type": "string"}, "quantity": {"type": "number", "default": 1} } } } }
        },
        "responses": {
          "200": { "description": "Updated cart" },
          "409": { "description": "Out of stock" }
        }
      }
    },
    "/api/cart/update": {
      "post": {
        "summary": "Update cart item quantity",
        "tags": ["Cart"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["productId", "variantId", "quantity"], "properties": { "productId": {"type": "string"}, "variantId": {"type": "string"}, "quantity": {"type": "number"} } } } } },
        "responses": { "200": { "description": "Updated cart" }, "409": { "description": "Out of stock" } }
      }
    },
    "/api/cart/remove": {
      "post": {
        "summary": "Remove item from cart",
        "tags": ["Cart"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["productId", "variantId"], "properties": { "productId": {"type": "string"}, "variantId": {"type": "string"} } } } } },
        "responses": { "200": { "description": "Updated cart" } }
      }
    },
...
  }
}